refactor(probe): extract agentic probes into probe_agentic module
Deduplicate cursor/codex file-walk assertion logic, table-drive sync probe
dispatch, and route opencode reads through sessions_in_probe_window so
probe.rs stays under the 1k-line threshold.
