+++ crates/hippo-daemon/src/lib.rs
@@ -16,6 +16,7 @@ pub mod metrics;
 pub mod native_messaging;
 pub mod opencode_session;
 pub mod probe;
+mod probe_agentic;
 #[cfg(feature = "otel")]
 pub mod process_metrics;
 pub mod schema_handshake;
